chore: refactoring import type + removing useless property config bean in PropertiesConfiguration
