+++ lsadf_core/src/main/java/com/lsadf/core/infra/config/ApplicationListenerConfiguration.java
@@ -15,9 +15,9 @@
  */
 package com.lsadf.core.infra.config;
 
-import com.lsadf.core.infra.cache.flush.CacheFlushService;
 import com.lsadf.core.infra.logging.ConfigurationLogger;
 import com.lsadf.core.infra.logging.properties.ConfigurationDisplayProperties;
+import com.lsadf.core.infra.valkey.cache.flush.CacheFlushService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.core.env.ConfigurableEnvironment;
