+++ lsadf_core/src/main/java/com/lsadf/core/infra/valkey/cache/flush/CacheFlushService.java
@@ -13,27 +13,18 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.lsadf.core.infra.cache.service;
+package com.lsadf.core.infra.valkey.cache.flush;
 
-public interface CacheService {
-
+public interface CacheFlushService {
   /**
-   * Check if the cache is enabled
-   *
-   * @return true if cache is enabled, false otherwise
+   * Flush the characteristics cache, and persists the characteristics of every entries in the
+   * database
    */
-  Boolean isEnabled();
+  void flushCharacteristics();
 
-  /** Toggle the cache enabling */
-  void toggleCacheEnabling();
+  /** Flush the currency cache, and persists the currency of every entries in the database */
+  void flushCurrencies();
 
-  /** Clear all the caches */
-  void clearCaches();
-
-  /**
-   * Clears the cache entry associated with the specified key.
-   *
-   * @param key the key for the cache entry to be cleared
-   */
-  void clearGameSaveValues(String key);
+  /** Flush the stage cache, and persists the stage of every entries in the database */
+  void flushStages();
 }
